[fixup] Move formatting functions to a utils file
