+++ src/services/scenario/ScenarioService.js
@@ -2,10 +2,12 @@
 // Licensed under the MIT license.
 
 import { CosmotechApiService } from '../../configs/Api.config';
-import rfdc from 'rfdc';
+import {
+  formatScenarioParametersForApi,
+  formatScenarioParametersFromApi
+} from '../../utils/ApiUtils';
 
 const ScenarioApi = new CosmotechApiService.ScenarioApi();
-const clone = rfdc();
 
 function findAllScenarios (organizationId, workspaceId) {
   return new Promise((resolve) => {
@@ -36,35 +38,6 @@ function createScenario (organizationId, workspaceId, scenario) {
     ScenarioApi.createScenario(organizationId, workspaceId, scenario, (error, data, response) => {
       resolve({ error, data, response });
     });
-  });
-}
-
-function formatScenarioParametersForApi (scenarioParameters) {
-  // Reformat scenario parameters to match the API expected types
-  return scenarioParameters.map(param => {
-    // Clone the original parameter
-    const newParam = clone(param);
-    // Cast boolean values into string values
-    if (newParam.varType === 'bool') {
-      newParam.value = newParam.value.toString();
-    }
-    return newParam;
-  });
-}
-
-function formatScenarioParametersFromApi (scenarioParameters) {
-  if (!scenarioParameters) {
-    return undefined;
-  }
-  // Reformat scenario parameters to match the front-end expected types
-  return scenarioParameters.map(param => {
-    // Clone the original parameter
-    const newParam = clone(param);
-    // Cast string values into boolean values
-    if (newParam.varType === 'bool') {
-      newParam.value = (newParam.value === 'true');
-    }
-    return newParam;
   });
 }
